Refactored updating thing logic into thing.h and thing.c
